Experimenting with header files
